email/requirements.txt
amqp==5.3.1
asgiref==3.9.2
async-timeout==5.0.1
billiard==4.2.2
boto3==1.40.42
botocore==1.40.42
celery==5.5.3
cffi==2.0.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
cryptography==46.0.1
Django==5.2.6
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-environ==0.12.0
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.1
jmespath==1.0.1
kombu==5.5.4
MarkupSafe==3.0.3
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycparser==2.23
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
redis==6.4.0
s3transfer==0.14.0
sendgrid==6.12.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.3
